refactor: modularize query and main window helpers
