+++ src/db/query_modules/__init__.py
@@ -1,0 +1,7 @@
+from .config_queries import *  # noqa: F403
+from .entity_queries import *  # noqa: F403
+from .history_queries import *  # noqa: F403
+from .library_queries import *  # noqa: F403
+from .track_queries import *  # noqa: F403
+
+__all__ = [name for name in globals() if not name.startswith("_")]
