+++ src/ui/main_window_parts/__init__.py
@@ -1,0 +1,10 @@
+from . import library_actions, library_filters, lyrics_actions, preferences
+from .lyrics_actions import canonical_lyrics_pair
+
+__all__ = [
+    "canonical_lyrics_pair",
+    "library_actions",
+    "library_filters",
+    "lyrics_actions",
+    "preferences",
+]
