+++ src/ui/widgets/lyrics_editor_parts/__init__.py
@@ -1,0 +1,4 @@
+from .flow_layout import FlowLayout
+from . import hotkeys
+
+__all__ = ["FlowLayout", "hotkeys"]
