+++ tests/test_migrations.py
@@ -183,7 +183,7 @@ class MigrationTests(unittest.TestCase):
             finally:
                 db.close()
 
-    def test_v3_database_upgrades_hotkey_bindings_column(self):
+    def test_v3_database_upgrades_hotkey_bindings_and_ui_state_columns(self):
         with tempfile.TemporaryDirectory() as tmp:
             db_path = Path(tmp) / "pylrcget.db.sqlite3"
             db = sqlite3.connect(str(db_path))
@@ -211,38 +211,11 @@ class MigrationTests(unittest.TestCase):
                 self.assertEqual(version, CURRENT_DB_VERSION)
                 columns = {row["name"] for row in db.execute("PRAGMA table_info(config_data)").fetchall()}
                 self.assertIn("hotkey_bindings_json", columns)
+                self.assertIn("ui_state_json", columns)
                 row = db.execute(
-                    "SELECT hotkey_bindings_json FROM config_data LIMIT 1"
+                    "SELECT hotkey_bindings_json, ui_state_json FROM config_data LIMIT 1"
                 ).fetchone()
                 self.assertEqual(row["hotkey_bindings_json"], "")
-            finally:
-                db.close()
-
-    def test_v4_database_upgrades_ui_state_column(self):
-        with tempfile.TemporaryDirectory() as tmp:
-            db_path = Path(tmp) / "pylrcget.db.sqlite3"
-            db = sqlite3.connect(str(db_path))
-            db.row_factory = sqlite3.Row
-            try:
-                db.execute(
-                    """
-                    CREATE TABLE config_data (
-                        id INTEGER PRIMARY KEY,
-                        hotkey_bindings_json TEXT DEFAULT ''
-                    )
-                    """
-                )
-                db.execute("INSERT INTO config_data (id) VALUES (1)")
-                db.execute("PRAGMA user_version=4")
-                db.commit()
-
-                upgrade_database_if_needed(db, 4)
-
-                version = int(db.execute("PRAGMA user_version").fetchone()[0])
-                self.assertEqual(version, CURRENT_DB_VERSION)
-                columns = {row["name"] for row in db.execute("PRAGMA table_info(config_data)").fetchall()}
-                self.assertIn("ui_state_json", columns)
-                row = db.execute("SELECT ui_state_json FROM config_data LIMIT 1").fetchone()
                 self.assertEqual(row["ui_state_json"], "")
             finally:
                 db.close()
